+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 * [breaking] Removed WarningFilter.fromConfig. ProjectConfig#initializeAnalyzer
   now does this from the other side (and cyclic deps are not a great idea).
+* Add the ability to cancel an analysis by passing in a CancelToken, which can
+  signal that the request has been cancelled. This can be useful in saving work
+  if you know that the result of the analysis won't be used.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.18] - 2018-03-27
@@ -454,7 +457,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 * Fix a class of race conditions and cache invalidation errors that can occur when there are concurrent analysis runs and edits to files.
-
 
 ## [2.0.0-alpha.19] - 2016-12-12